[AIP-21] Move operators.python_operator operators.python
