+++ airflow/contrib/example_dags/example_kubernetes_executor.py
@@ -22,7 +22,7 @@ This is an example dag for using the Kubernetes Executor.
 import os
 
 from airflow.models import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from airflow.utils.dates import days_ago
 
 args = {
