+++ airflow/contrib/example_dags/example_kubernetes_executor_config.py
@@ -23,7 +23,7 @@ import os
 
 from airflow.contrib.example_dags.libs.helper import print_stuff
 from airflow.models import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from airflow.utils.dates import days_ago
 
 default_args = {
