+++ airflow/example_dags/example_branch_python_dop_operator_3.py
@@ -24,7 +24,7 @@ or skipped on alternating runs.
 
 from airflow.models import DAG
 from airflow.operators.dummy_operator import DummyOperator
-from airflow.operators.python_operator import BranchPythonOperator
+from airflow.operators.python import BranchPythonOperator
 from airflow.utils.dates import days_ago
 
 args = {
