+++ airflow/example_dags/example_passing_params_via_test_command.py
@@ -23,7 +23,7 @@ from datetime import timedelta
 
 from airflow import DAG
 from airflow.operators.bash import BashOperator
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from airflow.utils.dates import days_ago
 
 dag = DAG(
