+++ airflow/example_dags/example_python_operator.py
@@ -23,7 +23,7 @@ import time
 from pprint import pprint
 
 from airflow.models import DAG
-from airflow.operators.python_operator import PythonOperator, PythonVirtualenvOperator
+from airflow.operators.python import PythonOperator, PythonVirtualenvOperator
 from airflow.utils.dates import days_ago
 
 args = {
