+++ airflow/example_dags/example_short_circuit_operator.py
@@ -21,7 +21,7 @@
 from airflow.models import DAG
 from airflow.models.baseoperator import chain
 from airflow.operators.dummy_operator import DummyOperator
-from airflow.operators.python_operator import ShortCircuitOperator
+from airflow.operators.python import ShortCircuitOperator
 from airflow.utils import dates
 
 args = {
