+++ airflow/example_dags/example_trigger_target_dag.py
@@ -25,7 +25,7 @@ Example usage of the TriggerDagRunOperator. This example holds 2 DAGs:
 
 from airflow.models import DAG
 from airflow.operators.bash import BashOperator
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from airflow.utils.dates import days_ago
 
 dag = DAG(
