+++ airflow/example_dags/example_xcom.py
@@ -20,7 +20,7 @@
 """Example DAG demonstrating the usage of XComs."""
 
 from airflow import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from airflow.utils.dates import days_ago
 
 args = {
