+++ tests/dags/test_cli_triggered_dags.py
@@ -21,7 +21,7 @@
 from datetime import timedelta
 
 from airflow.models import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from airflow.utils.timezone import datetime
 
 DEFAULT_DATE = datetime(2016, 1, 1)
