+++ tests/dags/test_impersonation_subdag.py
@@ -21,7 +21,7 @@ from datetime import datetime
 
 from airflow.models import DAG
 from airflow.operators.bash import BashOperator
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from airflow.operators.subdag_operator import SubDagOperator
 
 DEFAULT_DATE = datetime(2016, 1, 1)
