+++ tests/dags/test_mark_success.py
@@ -20,7 +20,7 @@ from datetime import datetime
 from time import sleep
 
 from airflow.models import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 
 DEFAULT_DATE = datetime(2016, 1, 1)
