+++ tests/dags/test_task_view_type_check.py
@@ -26,7 +26,7 @@ import logging
 from datetime import datetime
 
 from airflow.models import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 
 DEFAULT_DATE = datetime(2016, 1, 1)
 default_args = dict(
