+++ tests/dags_corrupted/test_impersonation_custom.py
@@ -29,7 +29,7 @@ from datetime import datetime
 from fake_datetime import FakeDatetime
 
 from airflow.models import DAG
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 
 DEFAULT_DATE = datetime(2016, 1, 1)
