+++ tests/utils/test_dot_renderer.py
@@ -24,7 +24,7 @@ import mock
 
 from airflow import DAG
 from airflow.operators.bash import BashOperator
-from airflow.operators.python_operator import PythonOperator
+from airflow.operators.python import PythonOperator
 from airflow.utils import dot_renderer
 
 START_DATE = datetime.datetime.now()
